don't be so quick to cast macro arguments to unsigned; if the arguments are really off_t's (e.g. to btodb), then you can lose information.  This was the "> 4G file systems don't work" bug; physio uses btodb, which was broken.

+++ sys/arch/amiga/include/param.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: param.h,v 1.17 1995/02/12 19:19:38 chopps Exp $	*/
+/*	$NetBSD: param.h,v 1.18 1995/03/18 07:22:51 cgd Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -126,9 +126,9 @@
 #define	btoc(x)	(((unsigned)(x)+(NBPG-1))>>PGSHIFT)
 
 #define	btodb(bytes)	 		/* calculates (bytes / DEV_BSIZE) */ \
-	((unsigned)(bytes) >> DEV_BSHIFT)
+	((bytes) >> DEV_BSHIFT)
 #define	dbtob(db)			/* calculates (db * DEV_BSIZE) */ \
-	((unsigned)(db) << DEV_BSHIFT)
+	((db) << DEV_BSHIFT)
 
 /*
  * Map a ``block device block'' to a file system block.
